version3.0

+++ src/main/java/pku/KeyValue.java
@@ -21,7 +21,7 @@ public interface KeyValue {
 
     public KeyValue put(String key, String value);
 
-    public short getInt(String key);
+    public int getInt(String key);
 
     public long getLong(String key);
